requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
aiosqlite>=0.19.0
httpx>=0.27.0
pydantic>=2.0
python-dotenv>=1.0.0
